+++ src/main/groovy/com/codebrig/arthur/observe/structure/filter/operator/relational/compare/IsEqualOperatorFilter.groovy
@@ -1,4 +1,4 @@
-package com.codebrig.arthur.observe.structure.filter.operator.relational
+package com.codebrig.arthur.observe.structure.filter.operator.relational.compare
 
 import com.codebrig.arthur.SourceNode
 import com.codebrig.arthur.observe.structure.StructureFilter
@@ -7,20 +7,20 @@ import com.codebrig.arthur.observe.structure.filter.RoleFilter
 import com.codebrig.arthur.observe.structure.filter.TypeFilter
 
 /**
- * Match by not equal operator
+ * Match by is equal operator
  *
  * @version 0.4
  * @since 0.3
  * @author <a href="mailto:brandon.fergerson@codebrig.com">Brandon Fergerson</a>
  */
-class IsNotEqualOperatorFilter extends StructureFilter<IsNotEqualOperatorFilter, Void> {
+class IsEqualOperatorFilter extends StructureFilter<IsEqualOperatorFilter, Void> {
 
     private final MultiFilter filter
 
-    IsNotEqualOperatorFilter() {
+    IsEqualOperatorFilter() {
         filter = MultiFilter.matchAll(
-                new RoleFilter("NOT"), new RoleFilter("EQUAL"),
-                new RoleFilter("OPERATOR"), new RoleFilter("RELATIONAL"),
+                new RoleFilter("EQUAL"), new RoleFilter("OPERATOR"),
+                new RoleFilter("RELATIONAL"),
                 new TypeFilter().reject("InfixExpression", "BinaryExpression")
         )
     }
