+++ src/main/groovy/com/codebrig/arthur/observe/structure/filter/operator/relational/compare/IsEqualTypeOperatorFilter.groovy
@@ -1,4 +1,4 @@
-package com.codebrig.arthur.observe.structure.filter.operator.relational
+package com.codebrig.arthur.observe.structure.filter.operator.relational.compare
 
 import com.codebrig.arthur.SourceNode
 import com.codebrig.arthur.observe.structure.StructureFilter
@@ -7,20 +7,21 @@ import com.codebrig.arthur.observe.structure.filter.RoleFilter
 import com.codebrig.arthur.observe.structure.filter.TypeFilter
 
 /**
- * Match by equal operator
+ * Match by is equal type operator
  *
  * @version 0.4
  * @since 0.3
  * @author <a href="mailto:brandon.fergerson@codebrig.com">Brandon Fergerson</a>
  */
-class IsEqualOperatorFilter extends StructureFilter<IsEqualOperatorFilter, Void> {
+class IsEqualTypeOperatorFilter extends StructureFilter<IsEqualTypeOperatorFilter, Void> {
 
     private final MultiFilter filter
 
-    IsEqualOperatorFilter() {
+    IsEqualTypeOperatorFilter() {
         filter = MultiFilter.matchAll(
-                new RoleFilter("EQUAL"), new RoleFilter("OPERATOR"),
-                new RoleFilter("RELATIONAL"),
+                new RoleFilter("IDENTICAL"), new RoleFilter("OPERATOR"),
+                new RoleFilter("RELATIONAL"), new RoleFilter("EXPRESSION"),
+                new RoleFilter("BINARY"),
                 new TypeFilter().reject("InfixExpression", "BinaryExpression")
         )
     }
