+++ src/main/groovy/com/codebrig/arthur/observe/structure/filter/operator/relational/compare/IsNotEqualOperatorFilter.groovy
@@ -1,4 +1,4 @@
-package com.codebrig.arthur.observe.structure.filter.operator.relational
+package com.codebrig.arthur.observe.structure.filter.operator.relational.compare
 
 import com.codebrig.arthur.SourceNode
 import com.codebrig.arthur.observe.structure.StructureFilter
@@ -7,21 +7,20 @@ import com.codebrig.arthur.observe.structure.filter.RoleFilter
 import com.codebrig.arthur.observe.structure.filter.TypeFilter
 
 /**
- * Match by not equal type operator
+ * Match by is not equal operator
  *
  * @version 0.4
  * @since 0.3
  * @author <a href="mailto:brandon.fergerson@codebrig.com">Brandon Fergerson</a>
  */
-class IsNotEqualTypeOperatorFilter extends StructureFilter<IsNotEqualTypeOperatorFilter, Void> {
+class IsNotEqualOperatorFilter extends StructureFilter<IsNotEqualOperatorFilter, Void> {
 
     private final MultiFilter filter
 
-    IsNotEqualTypeOperatorFilter() {
+    IsNotEqualOperatorFilter() {
         filter = MultiFilter.matchAll(
-                new RoleFilter("NOT"), new RoleFilter("IDENTICAL"),
+                new RoleFilter("NOT"), new RoleFilter("EQUAL"),
                 new RoleFilter("OPERATOR"), new RoleFilter("RELATIONAL"),
-                new RoleFilter("EXPRESSION"), new RoleFilter("BINARY"),
                 new TypeFilter().reject("InfixExpression", "BinaryExpression")
         )
     }
