+++ src/main/groovy/com/codebrig/arthur/observe/structure/filter/operator/relational/compare/IsNotEqualTypeOperatorFilter.groovy
@@ -1,4 +1,4 @@
-package com.codebrig.arthur.observe.structure.filter.operator.relational
+package com.codebrig.arthur.observe.structure.filter.operator.relational.compare
 
 import com.codebrig.arthur.SourceNode
 import com.codebrig.arthur.observe.structure.StructureFilter
@@ -7,21 +7,21 @@ import com.codebrig.arthur.observe.structure.filter.RoleFilter
 import com.codebrig.arthur.observe.structure.filter.TypeFilter
 
 /**
- * Match by equal type operator
+ * Match by is not equal type operator
  *
  * @version 0.4
  * @since 0.3
  * @author <a href="mailto:brandon.fergerson@codebrig.com">Brandon Fergerson</a>
  */
-class IsEqualTypeOperatorFilter extends StructureFilter<IsEqualTypeOperatorFilter, Void> {
+class IsNotEqualTypeOperatorFilter extends StructureFilter<IsNotEqualTypeOperatorFilter, Void> {
 
     private final MultiFilter filter
 
-    IsEqualTypeOperatorFilter() {
+    IsNotEqualTypeOperatorFilter() {
         filter = MultiFilter.matchAll(
-                new RoleFilter("IDENTICAL"), new RoleFilter("OPERATOR"),
-                new RoleFilter("RELATIONAL"), new RoleFilter("EXPRESSION"),
-                new RoleFilter("BINARY"),
+                new RoleFilter("NOT"), new RoleFilter("IDENTICAL"),
+                new RoleFilter("OPERATOR"), new RoleFilter("RELATIONAL"),
+                new RoleFilter("EXPRESSION"), new RoleFilter("BINARY"),
                 new TypeFilter().reject("InfixExpression", "BinaryExpression")
         )
     }
