+++ src/test/groovy/com/codebrig/arthur/observe/structure/filter/operator/relational/compare/IsEqualOperatorFilterTest.groovy
@@ -1,4 +1,4 @@
-package com.codebrig.arthur.observe.structure.filter.operator.relational
+package com.codebrig.arthur.observe.structure.filter.operator.relational.compare
 
 import com.codebrig.arthur.ArthurTest
 import com.codebrig.arthur.SourceLanguage
